third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-invalid-inputs-except-one-expected.txt

CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE ERROR: Dropped srcset candidate "foo.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE ERROR: Dropped srcset candidate "foo.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE ERROR: Dropped srcset candidate "foo.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has multiple 'h' descriptors or a mix of 'x' and 'h' descriptors.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since it has an unknown descriptor.
CONSOLE ERROR: Dropped srcset candidate "foo.jpg"
CONSOLE ERROR: Failed parsing 'srcset' attribute value since its 'x' descriptor is invalid.
CONSOLE ERROR: Dropped srcset candidate "bar.jpg"
PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("foo").clientWidth==200 is true
This test passes if the img tag below is a green square regardless of the scale factor. It ensures that invalid inputs are ignored and well-formed images are chosen regardless of their qualifiers, once they are the only candidate left
